Derek/requirements.txt
fastapi==0.112.1
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
mysql-connector-python==8.0.33
